backend/requirements.txt
flask
flask-smorest
marshmallow
flask-sqlalchemy
flask-migrate
psycopg2-binary
flask-cors
bcrypt
PyJWT
requests
pytest
